dodano promjena

+++ app/Http/Livewire/EmployeeShow.php
@@ -30,14 +30,21 @@ class EmployeeShow extends Component
     }
  
     public function saveEmployee()
-    {
+{
+    // Check if the current user is an admin
+    if (auth()->user()->is_admin) {
         $validatedData = $this->validate();
- 
+
         Employee::create($validatedData);
-        session()->flash('message','Employee Added Successfully');
+        session()->flash('message', 'Employee Added Successfully');
         $this->resetInput();
         $this->dispatchBrowserEvent('close-modal');
+    } else {
+        // User is not authorized, show forbidden popup or handle the situation accordingly
+        // For example, you can display an error message or redirect back
+        session()->flash('error', 'You are not authorized to perform this action.');
     }
+}
      
     public function editEmployee(int $employee_id)
     {
@@ -54,18 +61,25 @@ class EmployeeShow extends Component
     }
  
     public function updateEmployee()
-    {
+{
+    // Check if the current user is an admin
+    if (auth()->user()->is_admin) {
         $validatedData = $this->validate();
- 
-        Employee::where('id',$this->employee_id)->update([
+
+        Employee::where('id', $this->employee_id)->update([
             'nationalidnumber' => $validatedData['nationalidnumber'],
             'title' => $validatedData['title'],
             'birthdate' => $validatedData['birthdate']
         ]);
-        session()->flash('message','Employee Updated Successfully');
+        session()->flash('message', 'Employee Updated Successfully');
         $this->resetInput();
         $this->dispatchBrowserEvent('close-modal');
+    } else {
+        // User is not authorized, show forbidden popup or handle the situation accordingly
+        // For example, you can display an error message or redirect back
+        session()->flash('error', 'You are not authorized to perform this action.');
     }
+}
      
     public function deleteEmployee(int $employee_id)
     {
@@ -73,11 +87,17 @@ class EmployeeShow extends Component
     }
  
     public function destroyEmployee()
-    {
+{
+    // Check if the current user is an admin
+    if (auth()->user()->is_admin) {
         Employee::find($this->employee_id)->delete();
-        session()->flash('message','Employee Deleted Successfully');
+        session()->flash('message', 'Employee Deleted Successfully');
         $this->dispatchBrowserEvent('close-modal');
+    } else {
+        // User is not authorized, show forbidden message
+        $this->addError('authorization', 'You are not authorized to perform this action.');
     }
+}
  
     public function closeModal()
     {
