+++ resources/views/livewire/customer-show.blade.php
@@ -1,23 +1,30 @@
 <div>
-
     @include('livewire.customermodal')
 
     <div class="container">
-    <a href="{{ url('/employees') }}" class="btn btn-primary">{{ __('Employees') }}</a>
-    <a href="{{ url('/contacts') }}" class="btn btn-primary">{{ __('Contacts') }}</a>
+        <a href="{{ url('/employees') }}" class="btn btn-primary">{{ ('Employees') }}</a>
+        <a href="{{ url('/contacts') }}" class="btn btn-primary">{{ ('Contacts') }}</a>
         <div class="row">
             <div class="col-md-12">
                 @if (session()->has('message'))
                     <h5 class="alert alert-success">{{ session('message') }}</h5>
                 @endif
 
+                @if ($errors->has('authorization'))
+                    <div class="alert alert-danger">
+                        {{ $errors->first('authorization') }}
+                    </div>
+                @endif
+
                 <div class="card">
                     <div class="card-header">
                         <h4>Customer CRUD
                             <input type="search" wire:model="search" class="form-control float-end mx-2" placeholder="Search..." style="width: 230px" />
-                            <button type="button" class="btn btn-primary float-end" data-bs-toggle="modal" data-bs-target="#customerModal">
-                                Add New Customer
-                            </button>
+                            @if (auth()->user()->is_admin)
+                                <button type="button" class="btn btn-primary float-end" data-bs-toggle="modal" data-bs-target="#customerModal">
+                                    Add New Customer
+                                </button>
+                            @endif
                         </h4>
                     </div>
                     <div class="card-body">
@@ -39,10 +46,17 @@
                                         <td>{{ $customer->Email }}</td>
                                         <td>{{ $customer->City }}</td>
                                         <td>
-                                            <button type="button" data-bs-toggle="modal" data-bs-target="#updateCustomerModal" wire:click="editCustomer({{$customer->id}})" class="btn btn-primary">
-                                                Edit
-                                            </button>
-                                            <button type="button" data-bs-toggle="modal" data-bs-target="#deleteCustomerModal" wire:click="deleteCustomer({{$customer->id}})" class="btn btn-danger">Delete</button>
+                                            @if (auth()->user()->is_admin)
+                                                <button type="button" data-bs-toggle="modal" data-bs-target="#updateCustomerModal" wire:click="editCustomer({{$customer->id}})" class="btn btn-primary">
+                                                    Edit
+                                                </button>
+                                                <button type="button" data-bs-toggle="modal" data-bs-target="#deleteCustomerModal" wire:click="deleteCustomer({{$customer->id}})" class="btn btn-danger">Delete</button>
+                                            @else
+                                                <button type="button" class="btn btn-primary" disabled>
+                                                    Edit
+                                                </button>
+                                                <button type="button" class="btn btn-danger" disabled>Delete</button>
+                                            @endif
                                         </td>
                                     </tr>
                                 @empty
@@ -60,5 +74,4 @@
             </div>
         </div>
     </div>
-
 </div>
